requirements.txt
fastapi[standard]
SQLAlchemy
python-dotenv
flashrank
sentence-transformers